Extract reference analyzer
